project-root/ml_service/requirements.txt
flask
torch
torchvision
numpy
pyyaml
pillow
matplotlib